view data to line chart

+++ app/src/main/java/com/example/datn/data/model/HumiGraphModel.kt
@@ -2,5 +2,5 @@ package com.example.datn.data.model
 
 data class HumiGraphModel(
     var time: Int? = null,
-    var humi: Float? = null
+    var humi: Double? = null
 )
