+++ app/src/main/java/com/example/datn/data/model/TemGraphModel.kt
@@ -2,5 +2,5 @@ package com.example.datn.data.model
 
 data class TemGraphModel(
     var time: Int? = null,
-    var tem: Float? = null
+    var tem: Double? = null
 )
